Adjusted Device Object type decoding to use uint16_t rather than int.  Corrected unit tests.

+++ bacnet-stack/src/ihave.c
@@ -77,7 +77,7 @@ int ihave_decode_service_request(
     int len = 0;
     uint8_t tag_number = 0;
     uint32_t len_value = 0;
-    int decoded_type = 0;       /* for decoding */
+    uint16_t decoded_type = 0;       /* for decoding */
 
     if (apdu_len && data) {
         /* deviceIdentifier */
